+++ tests/w3d/cases/test_export.py
@@ -3,7 +3,7 @@
 
 import bpy
 from tests.utils import TestCase
-from io_mesh_w3d.export_w3d import save
+from io_mesh_w3d.w3d.export_w3d import save
 
 
 class TestExport(TestCase):
